pyproject.toml
[project]
name = "pr-ex3-kws"
version = "0.1.0"
description = "Pattern Recognition Exercise 3 – Keyword Spotting with DTW"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pillow",
    "matplotlib",
    "jupyter",
    "opencv-python>=4.12.0.88",
]
